Refactoring tests to separate out types
