+++ tests/test-batch.php
@@ -166,239 +166,6 @@ class BatchTest extends WP_UnitTestCase {
 	}
 
 	/**
-	 * Test that post batch gets run.
-	 */
-	public function test_post_finished_run() {
-		$posts = array();
-
-		// Create 5 posts.
-		for ( $x = 0; $x < 5; $x++ ) {
-			$posts[] = $this->factory->post->create();
-		}
-
-		$post_batch = new Posts();
-
-		$post_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'post',
-			'callback' => __NAMESPACE__ . '\\my_callback_function_test',
-			'args'     => array(
-				'posts_per_page' => 10,
-				'post_type'      => 'post',
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $post_batch->slug );
-
-		$this->assertFalse( $batch_status );
-
-		$run = $post_batch->run( 1 );
-
-		$batch_status = get_option( 'loco_batch_' . $post_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $posts as $post ) {
-			$meta = get_post_meta( $post, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_post_meta( $post, $post_batch->slug . '_status', true );
-			$this->assertEquals( 'success', $status );
-		}
-
-		// Run again so it skips some.
-		$run = $post_batch->run( 1 );
-	}
-
-	/**
-	 * Test that user batch gets run.
-	 */
-	public function test_user_finished_run() {
-		$users = $this->factory->user->create_many( 5 );
-
-		$user_batch = new Users();
-
-		$user_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'user',
-			'callback' => __NAMESPACE__ . '\\my_user_callback_function_test',
-			'args'     => array(
-				'number' => 10,
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $user_batch->slug );
-		$this->assertFalse( $batch_status );
-
-		$run = $user_batch->run( 1 );
-
-		$batch_status = get_option( 'loco_batch_' . $user_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $users as $user ) {
-			$meta = get_user_meta( $user, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_user_meta( $user, $user_batch->slug . '_status', true );
-			$this->assertEquals( 'success', $status );
-		}
-
-		// Run again so it skips some.
-		$run = $user_batch->run( 1 );
-	}
-
-	/**
-	 * Test that term batch gets run.
-	 */
-	public function test_term_finished_run() {
-		$terms = $this->factory->category->create_many( 5 );
-
-		$term_batch = new Terms();
-
-		$term_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'term',
-			'callback' => __NAMESPACE__ . '\\my_term_callback_function_test',
-			'args'     => array(
-				'number' => 10,
-				'taxonomy' => 'category',
-				'hide_empty' => false,
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $term_batch->slug );
-		$this->assertFalse( $batch_status );
-
-		$run = $term_batch->run( 1 );
-
-		$batch_status = get_option( 'loco_batch_' . $term_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $terms as $term ) {
-			$meta = get_term_meta( $term, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_term_meta( $term, $term_batch->slug . '_status', true );
-			$this->assertEquals( 'success', $status );
-		}
-
-		// Run again so it skips some.
-		$run = $term_batch->run( 1 );
-	}
-
-	/**
-	 * Test that you can clear individual result status.
-	 */
-	public function test_clear_result_status() {
-		$posts = array();
-
-		// Create 5 posts.
-		for ( $x = 0; $x < 5; $x++ ) {
-			$posts[] = $this->factory->post->create();
-		}
-
-		$post_batch = new Posts();
-
-		$post_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'post',
-			'callback' => __NAMESPACE__ . '\\my_callback_function_test',
-			'args'     => array(
-				'posts_per_page' => 10,
-				'post_type'      => 'post',
-			),
-		) );
-
-		$run = $post_batch->run( 1 );
-
-		$post_batch->clear_result_status();
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $posts as $post ) {
-			$meta = get_post_meta( $post, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_post_meta( $post, $post_batch->slug . '_status', true );
-			$this->assertEquals( '', $status );
-		}
-
-		$batches = locomotive_get_all_batches();
-		$this->assertEquals( 'reset', $batches['hey-there']['status'] );
-	}
-
-	/**
-	 * Test that you can clear individual result status.
-	 */
-	public function test_clear_user_result_status() {
-		$users = $this->factory->user->create_many( 5 );
-
-		$user_batch = new Users();
-
-		$user_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'user',
-			'callback' => __NAMESPACE__ . '\\my_user_callback_function_test',
-			'args'     => array(
-				'number' => 7,
-			),
-		) );
-
-		$run = $user_batch->run( 1 );
-
-		$user_batch->clear_result_status();
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $users as $user ) {
-			$meta = get_user_meta( $user, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_user_meta( $user, $user_batch->slug . '_status', true );
-			$this->assertEquals( '', $status );
-		}
-
-		$batches = locomotive_get_all_batches();
-		$this->assertEquals( 'reset', $batches['hey-there']['status'] );
-	}
-
-	/**
-	 * Test that you can clear individual result status.
-	 */
-	public function test_clear_term_result_status() {
-		$terms = $this->factory->tag->create_many( 5 );
-
-		$term_batch = new Terms();
-
-		$term_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'term',
-			'callback' => __NAMESPACE__ . '\\my_term_callback_function_test',
-			'args'     => array(
-				'number' => 7,
-				'taxonomy' => 'post_tag',
-				'hide_empty' => false,
-			),
-		) );
-
-		$run = $term_batch->run( 1 );
-
-		$term_batch->clear_result_status();
-
-		// Loop through each post and make sure our value was set.
-		foreach ( $terms as $term ) {
-			$meta = get_term_meta( $term, 'custom-key', true );
-			$this->assertEquals( 'my-value', $meta );
-
-			$status = get_term_meta( $term, $term_batch->slug . '_status', true );
-			$this->assertEquals( '', $status );
-		}
-
-		$batches = locomotive_get_all_batches();
-		$this->assertEquals( 'reset', $batches['hey-there']['status'] );
-	}
-
-	/**
 	 * Test that batch gets run.
 	 */
 	public function test_running_run() {
@@ -425,90 +192,7 @@ class BatchTest extends WP_UnitTestCase {
 		$run = $post_batch->run( 1 );
 
 		$batch_status = get_option( 'loco_batch_' . $post_batch->slug );
-		$this->assertEquals ( 'running', $batch_status['status'] );
-	}
-
-	/**
-	 * Test that batch gets run.
-	 */
-	public function test_offset_run() {
-		// Create 5 posts.
-		for ( $x = 0; $x < 13; $x++ ) {
-			$this->factory->post->create();
-		}
-
-		$post_batch = new Posts();
-		$post_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'post',
-			'callback' => 'my_callback_function_test',
-			'args'     => array(
-				'posts_per_page' => 10,
-				'post_type'      => 'post',
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $post_batch->slug );
-		$this->assertFalse( $batch_status );
-
-		$run = $post_batch->run( 2 );
-
-		$batch_status = get_option( 'loco_batch_' . $post_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
-	}
-
-	/**
-	 * Test that terms offset batch gets run.
-	 */
-	public function test_terms_offset_run() {
-		$users = $this->factory->category->create_many( 8 );
-
-		$term_batch = new Terms();
-		$term_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'term',
-			'callback' => 'my_callback_function_test',
-			'args'     => array(
-				'number' => 5,
-				'offset' => 5,
-				'taxonomy' => 'category',
-				'hide_empty' => false,
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $term_batch->slug );
-		$this->assertFalse( $batch_status );
-
-		$run = $term_batch->run( 2 );
-
-		$batch_status = get_option( 'loco_batch_' . $term_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
-	}
-
-	/**
-	 * Test that users offset batch gets run.
-	 */
-	public function test_users_offset_run() {
-		$users = $this->factory->user->create_many( 8 );
-
-		$user_batch = new Users();
-		$user_batch->register( array(
-			'name'     => 'Hey there',
-			'type'     => 'user',
-			'callback' => __NAMESPACE__ . '\\my_callback_function_test',
-			'args'     => array(
-				'number' => 5,
-				'offset' => 5,
-			),
-		) );
-
-		$batch_status = get_option( 'loco_batch_' . $user_batch->slug );
-		$this->assertFalse( $batch_status );
-
-		$run = $user_batch->run( 2 );
-
-		$batch_status = get_option( 'loco_batch_' . $user_batch->slug );
-		$this->assertEquals( 'finished', $batch_status['status'] );
+		$this->assertEquals( 'running', $batch_status['status'] );
 	}
 
 	/**
@@ -589,25 +273,8 @@ function my_callback_function_test( $result ) {
 /**
  * My callback function test.
  *
- * @param WP_Post $result Result item.
- */
-function my_user_callback_function_test( $result ) {
-	update_user_meta( $result->data->ID, 'custom-key', 'my-value' );
-}
-
-/**
- * My callback function test for terms.
- *
- * @param WP_Term $result Result item.
- */
-function my_term_callback_function_test( $result ) {
-	update_term_meta( $result->data->term_id, 'custom-key', 'my-value' );
-}
-
-/**
- * My callback function test.
- *
- * @throws Exception Not worknig.
+ * @param object $result Result item.
+ * @throws \Exception Not working.
  */
 function my_callback_function_test_false( $result ) {
 	throw new \Exception( 'Not working' );
